requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
accelerate=1.6.0=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.16=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
appnope=0.1.4=pyhd8ed1ab_1
asgiref=3.8.1=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
attrs=25.3.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.3.0=pypi_0
beautifulsoup4=4.13.4=pypi_0
build=1.2.2.post1=pypi_0
bzip2=1.0.8=h6c40b1e_6
ca-certificates=2025.1.31=hbd8a1cb_1
cachetools=5.5.2=pypi_0
cbor=1.0.0=pypi_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
chroma-hnswlib=0.7.6=pypi_0
chromadb=0.6.3=pypi_0
click=8.1.8=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pyhd8ed1ab_1
dataclasses-json=0.6.7=pypi_0
datasets=3.5.0=pypi_0
debugpy=1.8.14=py311hc356e98_0
decorator=5.2.1=pyhd8ed1ab_0
deprecated=1.2.18=pypi_0
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
durationpy=0.9=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
fastapi=0.115.9=pypi_0
fastjsonschema=2.21.1=pypi_0
filelock=3.18.0=pypi_0
filetype=1.2.0=pypi_0
flagembedding=1.3.4=pypi_0
flatbuffers=25.2.10=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.12.0=pypi_0
ftfy=6.3.1=pypi_0
google-auth=2.39.0=pypi_0
googleapis-common-protos=1.70.0=pypi_0
greenlet=3.1.1=pypi_0
grpcio=1.71.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.8=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.1=pypi_0
httpx-sse=0.4.0=pypi_0
huggingface-hub=0.30.2=pypi_0
humanfriendly=10.0=pypi_0
idna=3.10=pypi_0
ijson=3.3.0=pypi_0
importlib-metadata=8.6.1=pyha770c72_0
importlib-resources=6.5.2=pypi_0
inscriptis=2.6.0=pypi_0
ipykernel=6.29.5=pyh57ce528_0
ipython=9.1.0=pyhfb0248b_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
ir-datasets=0.5.10=pypi_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pypi_0
jiter=0.9.0=pypi_0
joblib=1.4.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-pydantic=0.6=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
jupytext=1.17.0=pypi_0
kubernetes=32.0.1=pypi_0
langchain=0.3.23=pypi_0
langchain-chroma=0.2.3=pypi_0
langchain-community=0.3.21=pypi_0
langchain-core=0.3.54=pypi_0
langchain-deepseek=0.1.3=pypi_0
langchain-experimental=0.3.4=pypi_0
langchain-mcp-adapters=0.0.7=pypi_0
langchain-mcp-tools=0.2.2=pypi_0
langchain-ollama=0.3.1=pypi_0
langchain-openai=0.3.12=pypi_0
langchain-text-splitters=0.3.8=pypi_0
langgraph=0.3.29=pypi_0
langgraph-checkpoint=2.0.24=pypi_0
langgraph-prebuilt=0.1.8=pypi_0
langgraph-sdk=0.1.61=pypi_0
langsmith=0.3.30=pypi_0
libcxx=20.1.3=hf95d169_0
libffi=3.4.4=hecd8cb5_1
libsodium=1.0.18=hbcb3906_1
lxml=5.3.2=pypi_0
lz4=4.4.4=pypi_0
markdown-it-py=3.0.0=pypi_0
marker-pdf=0.2.6=pypi_0
markupsafe=3.0.2=pypi_0
marshmallow=3.26.1=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mcp=1.6.0=pypi_0
mdit-py-plugins=0.4.2=pypi_0
mdurl=0.1.2=pypi_0
milvus=2.3.5=pypi_0
mmh3=5.1.0=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
multidict=6.4.3=pypi_0
multiprocess=0.70.16=pypi_0
mypy-extensions=1.0.0=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pypi_0
numpy=1.26.2=pypi_0
oauthlib=3.2.2=pypi_0
ollama=0.4.7=pypi_0
onnxruntime=1.21.0=pypi_0
openai=1.73.0=pypi_0
opencv-python=4.11.0.86=pypi_0
openssl=3.5.0=hc426f3f_0
opentelemetry-api=1.32.1=pypi_0
opentelemetry-exporter-otlp-proto-common=1.32.1=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.32.1=pypi_0
opentelemetry-instrumentation=0.53b1=pypi_0
opentelemetry-instrumentation-asgi=0.53b1=pypi_0
opentelemetry-instrumentation-fastapi=0.53b1=pypi_0
opentelemetry-proto=1.32.1=pypi_0
opentelemetry-sdk=1.32.1=pypi_0
opentelemetry-semantic-conventions=0.53b1=pypi_0
opentelemetry-util-http=0.53b1=pypi_0
orjson=3.10.16=pypi_0
ormsgpack=1.9.1=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pyhd8ed1ab_1
pdftext=0.3.20=pypi_0
peft=0.15.2=pypi_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=10.4.0=pypi_0
pip=25.0=py311hecd8cb5_0
platformdirs=4.3.7=pyh29332c3_0
posthog=3.25.0=pypi_0
prompt-toolkit=3.0.51=pyha770c72_0
propcache=0.3.1=pypi_0
protobuf=5.29.4=pypi_0
psutil=7.0.0=py311h4d7f069_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyarrow=19.0.1=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.2=pypi_0
pydantic=2.11.3=pypi_0
pydantic-core=2.33.1=pypi_0
pydantic-settings=2.8.1=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pyjson5=1.6.8=pypi_0
pypdf=5.4.0=pypi_0
pypdfium2=4.30.1=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.2.0=pypi_0
python=3.11.11=h4d6d9e5_0
python-dateutil=2.9.0.post0=pyhff2d567_1
python-dotenv=1.1.0=pypi_0
python_abi=3.11=2_cp311
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=24.0.1=py311habfacb3_1
rapidfuzz=3.13.0=pypi_0
ray=2.44.1=pypi_0
readline=8.2=hca72f7f_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=14.0.0=pypi_0
rpds-py=0.24.0=pypi_0
rsa=4.9=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
sentence-transformers=4.1.0=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=75.8.0=py311hecd8cb5_0
shellingham=1.5.4=pypi_0
six=1.17.0=pyhd8ed1ab_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlalchemy=2.0.40=pypi_0
sqlite=3.45.3=h6c40b1e_0
sse-starlette=2.2.1=pypi_0
stack_data=0.6.3=pyhd8ed1ab_1
starlette=0.45.3=pypi_0
surya-ocr=0.4.5=pypi_0
sympy=1.13.3=pypi_0
tabulate=0.9.0=pypi_0
tenacity=9.1.2=pypi_0
texify=0.1.10=pypi_0
threadpoolctl=3.6.0=pypi_0
tiktoken=0.9.0=pypi_0
tk=8.6.14=h4d00af3_0
tokenizers=0.19.1=pypi_0
torch=2.2.2=pypi_0
tornado=6.4.2=py311h4d7f069_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.44.2=pypi_0
trec-car-tools=2.6=pypi_0
typer=0.15.2=pypi_0
typing-inspect=0.9.0=pypi_0
typing-inspection=0.4.0=pypi_0
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025.2=pypi_0
unlzw3=0.2.3=pypi_0
urllib3=2.4.0=pypi_0
uvicorn=0.34.1=pypi_0
uvloop=0.21.0=pypi_0
warc3-wet=0.2.5=pypi_0
warc3-wet-clueweb09=0.2.5=pypi_0
watchfiles=1.0.5=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
websocket-client=1.8.0=pypi_0
websockets=15.0.1=pypi_0
wheel=0.45.1=py311hecd8cb5_0
wrapt=1.17.2=pypi_0
xxhash=3.5.0=pypi_0
xz=5.6.4=h46256e1_1
yarl=1.19.0=pypi_0
zeromq=4.3.5=h93d8f39_0
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h4b97444_1
zlib-state=0.1.9=pypi_0
zstandard=0.23.0=pypi_0
